+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Layout toggle buttons on the items list to switch between table and grid views.
+- Supplier and unit dropdown filters on the items list for more precise results.
 
 ### Removed